--- src/main/java/com/example/newsfeed/dto/request/PostSaveRequestDto.java
@@ -1,12 +1,0 @@
-package com.example.newsfeed.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class PostSaveRequestDto {
-
-    private String title;
-
-    private String description;
-
-}
