--- src/main/java/com/example/newsfeed/dto/request/PostUpdateRequestDto.java
@@ -1,15 +1,0 @@
-package com.example.newsfeed.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class PostUpdateRequestDto {
-
-    private Long postId;
-
-    private String title;
-
-    private String description;
-
-
-}
