--- src/main/java/com/example/newsfeed/exception/GlobalExceptionHandler.java
@@ -1,37 +1,0 @@
-package com.example.newsfeed.exception;
-
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(ResponseStatusException.class)
-    public ResponseEntity<Map<String, String>> handleIllegalArgument(ResponseStatusException ex) {
-        log.error("에러 발생: {}", ex.getMessage());
-
-        // 에러 메시지와 상태 코드 설정
-        Map<String, String> errorResponse = new HashMap<>();
-        errorResponse.put("status", ex.getStatusCode().toString());
-        errorResponse.put("message", ex.getReason());
-
-        return new ResponseEntity<>(errorResponse, ex.getStatusCode());
-    }
-
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<String> handleRuntimeException(RuntimeException ex) {
-        // 모든 예외에 대한 처리
-        log.error("알 수 없는 에러 발생: {}", ex.getMessage());
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error: " + ex.getMessage());
-    }
-}
