+++ src/main/java/com/example/newsfeed/post/dto/request/PostStateChangeRequestDto.java
@@ -1,0 +1,13 @@
+package com.example.newsfeed.post.dto.request;
+
+import com.example.newsfeed.post.entity.State;
+import lombok.Getter;
+
+@Getter
+public class PostStateChangeRequestDto {
+
+    private Long postId;
+
+    private State state;
+
+}
