+++ src/main/java/com/example/newsfeed/post/repository/PostRepository.java
@@ -1,0 +1,10 @@
+package com.example.newsfeed.post.repository;
+
+import com.example.newsfeed.post.entity.Post;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post,Long> {
+    Page<Post> findAllByOrderByCreatedAtDesc(Pageable pageable);
+}
