+++ src/main/java/com/example/newsfeed/post/util/Const.java
@@ -1,0 +1,5 @@
+package com.example.newsfeed.post.util;
+
+public abstract class Const {
+    public static final String LOGIN_USER = "LOGIN_USER";
+}
