+++ src/main/java/com/example/newsfeed/post/util/Filter/FriendFilter.java
@@ -1,0 +1,4 @@
+package com.example.newsfeed.post.util.Filter;
+
+public class FriendFilter {
+}
