+++ src/main/java/com/example/newsfeed/post/util/Filter/LoginFilter.java
@@ -1,0 +1,45 @@
+package com.example.newsfeed.post.util.Filter;
+
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.util.PatternMatchUtils;
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletRequest;
+import java.io.IOException;
+
+@Slf4j
+public class LoginFilter implements Filter {
+
+    private static final String[] WHITE_LIST = {"/posts"};
+
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
+            throws IOException, ServletException {
+
+        HttpServletRequest httpRequest  = (HttpServletRequest) request;
+        HttpServletResponse httpResponse = (HttpServletResponse) response;
+        String requestURI = httpRequest.getRequestURI();
+
+        // 화이트리스트에 포함되지 않은 경우, 로그인 여부 확인
+        if (!isWhiteList(requestURI)) {
+            HttpSession session = httpRequest.getSession(false);
+
+            if (session == null ||session.getAttribute("LOGIN_USER") == null) {
+                httpResponse.sendError(HttpServletResponse.SC_UNAUTHORIZED, "로그인 해주세요.");
+                return;
+            }
+        }
+        chain.doFilter(request, response);
+    }
+
+    private boolean isWhiteList(String requestURI) {
+        return PatternMatchUtils.simpleMatch(WHITE_LIST, requestURI);
+    }
+}
