--- src/main/java/com/example/newsfeed/util/Const.java
@@ -1,5 +1,0 @@
-package com.example.newsfeed.util;
-
-public abstract class Const {
-    public static final String LOGIN_USER = "LOGIN_USER";
-}
