--- src/main/java/com/example/newsfeed/util/Filter/FilterConfig.java
@@ -1,20 +1,0 @@
-package com.example.newsfeed.util.Filter;
-
-import jakarta.servlet.Filter;
-
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FilterConfig {
-
-    @Bean
-    public FilterRegistrationBean<Filter> loginFilter() {
-        FilterRegistrationBean<Filter> registrationBean = new FilterRegistrationBean<>();
-        registrationBean.setFilter(new LoginFilter());
-        registrationBean.setOrder(1);
-        registrationBean.addUrlPatterns("/*");
-        return registrationBean;
-    }
-}
