--- src/main/java/com/example/newsfeed/util/Filter/FriendFilter.java
@@ -1,4 +1,0 @@
-package com.example.newsfeed.util.Filter;
-
-public class FriendFilter {
-}
