+++ src/main/resources/application.properties
@@ -1,0 +1,16 @@
+spring.application.name=News-Feed
+
+spring.datasource.url=jdbc:mysql://localhost:3306/post
+spring.datasource.username=root
+spring.datasource.password=1234
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+spring.jpa.hibernate.ddl-auto=create
+
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.use_sql_comments=true
+
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
